requirements.txt
Django==3.1
django-allauth==0.41.0
celery==4.4.1
django-celery-beat==2.0.0
django-extensions==2.2.9
django-facebook==6.0.3
django-filter==2.2.0
django-multiselectfield==0.1.12
django-qsstats-magic==1.1.0
django-rest-auth==0.9.5
django-restful-admin==1.1.1
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-csv==2.1.0
PySocks==1.7.1
requests-html==0.10.0
torrequest==0.1.0
elasticsearch-dsl==7.3.0
youtube-search-python==1.3.1
django-computed-property==0.3.0
django-admin-rangefilter==0.6.3
django-admin-list-filter-dropdown==1.0.3
